requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-text-splitters
bs4
python-dotenv
